Refactoring impl package
* separate impl package with implmentation classes
* deleted NER service (req. too large menory)
