+++ src/main/java/edu/brandeis/cs/lappsgrid/clearnlp/impl/POSTagger.java
@@ -1,5 +1,6 @@
-package edu.brandeis.cs.lappsgrid.clearnlp;
+package edu.brandeis.cs.lappsgrid.clearnlp.impl;
 
+import edu.brandeis.cs.lappsgrid.clearnlp.AbstractClearNLPWebService;
 import edu.emory.clir.clearnlp.dependency.DEPNode;
 import edu.emory.clir.clearnlp.dependency.DEPTree;
 import org.lappsgrid.serialization.Data;
@@ -13,15 +14,16 @@ import java.util.List;
 import static org.lappsgrid.discriminator.Discriminators.Uri;
 
 @org.lappsgrid.annotations.ServiceMetadata(
-        description = "ClearNLP Tokenizer",
+        description = "ClearNLP Part-of-speech Tagger",
         requires_format = { "text", "lif" },
         produces_format = { "lif" },
-        produces = { "token" }
+        produces = { "pos" }
 )
-public class Tokenizer extends AbstractClearNLPWebService {
+public class POSTagger extends AbstractClearNLPWebService {
 
-    public Tokenizer() {
+    public POSTagger() {
         super();
+        getComponent(POS);
     }
 
     @Override
@@ -29,21 +31,24 @@ public class Tokenizer extends AbstractClearNLPWebService {
         String text = container.getText();
         View view = container.newView();
         String serviceName = this.getClass().getName();
-        view.addContains(Uri.TOKEN,
+        view.addContains(Uri.POS,
                 String.format("%s:%s", serviceName, getVersion()),
-                "tokenizer:clearnlp");
-        List<List<String>> sents = split(text);
+                "pos-tagger:clearnlp");
+        List<DEPTree> sents = process(text);
         List<int[]> spans = getSpans(text);
         int tokenSoFar = 0;
         int sid = 1;
-        for (List<String> sent : sents) {
-            int tid = 1;
-            for (String token : sent) {
+        for (DEPTree sent : sents) {
+            for (DEPNode token : sent) {
+                int tid = token.getID();
+                if (tid == 0) { continue; } // root node
                 int[] span = spans.get(tokenSoFar++);
                 int start = span[0]; int end = span[1];
                 Annotation tok = view.newAnnotation(
-                        makeID(TOKEN_ID, sid, tid++), Uri.TOKEN, start, end);
-                tok.addFeature("word", token);
+                        makeID(TOKEN_ID, sid, tid), Uri.TOKEN, start, end);
+                tok.addFeature("pos", token.getPOSTag());
+                tok.addFeature("word", token.getWordForm());
+
             }
             sid++;
         }
