+++ src/main/java/edu/brandeis/cs/lappsgrid/clearnlp/impl/Tokenizer.java
@@ -1,5 +1,6 @@
-package edu.brandeis.cs.lappsgrid.clearnlp;
+package edu.brandeis.cs.lappsgrid.clearnlp.impl;
 
+import edu.brandeis.cs.lappsgrid.clearnlp.AbstractClearNLPWebService;
 import edu.emory.clir.clearnlp.dependency.DEPNode;
 import edu.emory.clir.clearnlp.dependency.DEPTree;
 import org.lappsgrid.serialization.Data;
@@ -13,16 +14,15 @@ import java.util.List;
 import static org.lappsgrid.discriminator.Discriminators.Uri;
 
 @org.lappsgrid.annotations.ServiceMetadata(
-        description = "ClearNLP Part-of-speech Tagger",
+        description = "ClearNLP Tokenizer",
         requires_format = { "text", "lif" },
         produces_format = { "lif" },
-        produces = { "pos" }
+        produces = { "token" }
 )
-public class POSTagger extends AbstractClearNLPWebService {
+public class Tokenizer extends AbstractClearNLPWebService {
 
-    public POSTagger() {
+    public Tokenizer() {
         super();
-        getComponent(POS);
     }
 
     @Override
@@ -30,24 +30,21 @@ public class POSTagger extends AbstractClearNLPWebService {
         String text = container.getText();
         View view = container.newView();
         String serviceName = this.getClass().getName();
-        view.addContains(Uri.POS,
+        view.addContains(Uri.TOKEN,
                 String.format("%s:%s", serviceName, getVersion()),
-                "pos-tagger:clearnlp");
-        List<DEPTree> sents = process(text);
+                "tokenizer:clearnlp");
+        List<List<String>> sents = split(text);
         List<int[]> spans = getSpans(text);
         int tokenSoFar = 0;
         int sid = 1;
-        for (DEPTree sent : sents) {
-            for (DEPNode token : sent) {
-                int tid = token.getID();
-                if (tid == 0) { continue; } // root node
+        for (List<String> sent : sents) {
+            int tid = 1;
+            for (String token : sent) {
                 int[] span = spans.get(tokenSoFar++);
                 int start = span[0]; int end = span[1];
                 Annotation tok = view.newAnnotation(
-                        makeID(TOKEN_ID, sid, tid), Uri.TOKEN, start, end);
-                tok.addFeature("pos", token.getPOSTag());
-                tok.addFeature("word", token.getWordForm());
-
+                        makeID(TOKEN_ID, sid, tid++), Uri.TOKEN, start, end);
+                tok.addFeature("word", token);
             }
             sid++;
         }
