+++ src/main/webapp/WEB-INF/serviceimpl/DependencyParser.xml
@@ -3,7 +3,7 @@
 <beans>
   <bean id="target" class="jp.go.nict.langrid.servicecontainer.handler.TargetServiceFactory" >
     <property name="service">
-      <bean class="edu.brandeis.cs.lappsgrid.clearnlp.DependencyParser" />
+      <bean class="edu.brandeis.cs.lappsgrid.clearnlp.impl.DependencyParser" />
     </property>
   </bean>
 </beans>
