+++ src/main/webapp/WEB-INF/serviceimpl/Tokenizer.xml
@@ -3,7 +3,7 @@
 <beans>
   <bean id="target" class="jp.go.nict.langrid.servicecontainer.handler.TargetServiceFactory" >
     <property name="service">
-      <bean class="edu.brandeis.cs.lappsgrid.clearnlp.Tokenizer" />
+      <bean class="edu.brandeis.cs.lappsgrid.clearnlp.impl.Tokenizer" />
     </property>
   </bean>
 </beans>
